chore: extract composer methods from state.rs into state/composer.rs
